+++ src/amun/gpu/dl4mt/cellstate.cu
@@ -6,7 +6,7 @@ using namespace std;
 namespace amunmt {
 namespace GPU {
 
-std::string CellState::Debug(size_t verbosity) const
+std::string CellState::Debug(unsigned verbosity) const
 {
   std::stringstream strm;
